Add registration & authorization

+++ controllers/cards.js
@@ -1,5 +1,10 @@
 const Card = require('../models/card');
-const { STATUS_400, STATUS_404, STATUS_500 } = require('../utils/constants');
+const {
+  STATUS_403,
+  STATUS_400,
+  STATUS_404,
+  STATUS_500
+} = require('../utils/constants');
 
 const getCards = (req, res) => {
   Card.find({})
@@ -21,14 +26,42 @@ const createCard = (req, res) => {
     });
 };
 
+// const deleteCard = (req, res) => {
+//   Card.findByIdAndRemove(req.params.cardId)
+//     .then((card) => {
+//       if (!card) {
+//         res.status(STATUS_404).send({ message: 'Карточка с указанным _id не найдена' });
+//         return;
+//       }
+//       res.send(card);
+//     })
+//     .catch((err) => {
+//       if (err.name === 'CastError') {
+//         res.status(STATUS_400).send({ message: 'Переданы некорректные данные карточки' });
+//         return;
+//       }
+//       res.status(STATUS_500).send({ message: 'На сервере произошла ошибка' });
+//     });
+// };
+
 const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
         res.status(STATUS_404).send({ message: 'Карточка с указанным _id не найдена' });
         return;
       }
-      res.send(card);
+      if (card.owner.toString() !== req.user._id) {
+        res.status(STATUS_403).send({ message: 'У пользователя нет прав для удаления этой карточки' });
+        return;
+      }
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((deletedCard) => {
+          res.send(deletedCard);
+        })
+        .catch(() => {
+          res.status(STATUS_500).send({ message: 'На сервере произошла ошибка' });
+        });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
@@ -38,6 +71,8 @@ const deleteCard = (req, res) => {
       res.status(STATUS_500).send({ message: 'На сервере произошла ошибка' });
     });
 };
+
+// /////////////////////////////////////////////////
 
 const likeCard = (req, res) => {
   Card.findByIdAndUpdate(
